Move response scripts and tests
